+++ executor/invoker/src/commonMain/kotlin/io/github/charlietap/chasm/executor/invoker/ext/IndexExt.kt
@@ -1,5 +1,5 @@
 package io.github.charlietap.chasm.executor.invoker.ext
 
-import io.github.charlietap.chasm.ast.module.Index
+import io.github.charlietap.chasm.ir.module.Index
 
-inline fun Index.index() = idx.toInt()
+inline fun Index.index() = idx
