+++ executor/runtime-internal/src/commonMain/kotlin/io/github/charlietap/chasm/executor/runtime/execution/ExecutionContext.kt
@@ -1,19 +1,19 @@
 package io.github.charlietap.chasm.executor.runtime.execution
 
-import io.github.charlietap.chasm.ast.module.Index
-import io.github.charlietap.chasm.ast.type.DefinedType
-import io.github.charlietap.chasm.ast.type.SubType
 import io.github.charlietap.chasm.config.RuntimeConfig
 import io.github.charlietap.chasm.executor.runtime.instance.ModuleInstance
 import io.github.charlietap.chasm.executor.runtime.stack.ControlStack
 import io.github.charlietap.chasm.executor.runtime.stack.ValueStack
 import io.github.charlietap.chasm.executor.runtime.store.Store
-import io.github.charlietap.chasm.type.factory.DefinedTypeUnrollerFactory
-import io.github.charlietap.chasm.type.matching.DefinedTypeReverseLookup
-import io.github.charlietap.chasm.type.matching.TypeMatcherContext
-import io.github.charlietap.chasm.type.rolling.DefinedTypeUnroller
-import io.github.charlietap.chasm.type.rolling.substitution.ConcreteHeapTypeSubstitutor
-import io.github.charlietap.chasm.type.rolling.substitution.TypeIndexToDefinedTypeSubstitutor
+import io.github.charlietap.chasm.ir.module.Index
+import io.github.charlietap.chasm.ir.type.DefinedType
+import io.github.charlietap.chasm.ir.type.SubType
+import io.github.charlietap.chasm.type.ir.factory.DefinedTypeUnrollerFactory
+import io.github.charlietap.chasm.type.ir.matching.DefinedTypeReverseLookup
+import io.github.charlietap.chasm.type.ir.matching.TypeMatcherContext
+import io.github.charlietap.chasm.type.ir.rolling.DefinedTypeUnroller
+import io.github.charlietap.chasm.type.ir.rolling.substitution.ConcreteHeapTypeSubstitutor
+import io.github.charlietap.chasm.type.ir.rolling.substitution.TypeIndexToDefinedTypeSubstitutor
 
 data class ExecutionContext(
     val cstack: ControlStack,
@@ -25,7 +25,7 @@ data class ExecutionContext(
 ) : TypeMatcherContext {
 
     override val lookup: (Index.TypeIndex) -> DefinedType? = { index ->
-        instance.types.getOrNull(index.idx.toInt())
+        instance.types.getOrNull(index.idx)
     }
 
     override val reverseLookup: DefinedTypeReverseLookup
