+++ executor/runtime-internal/src/commonMain/kotlin/io/github/charlietap/chasm/executor/runtime/ext/StructTypeExt.kt
@@ -1,17 +1,17 @@
 package io.github.charlietap.chasm.executor.runtime.ext
 
-import io.github.charlietap.chasm.ast.module.Index
-import io.github.charlietap.chasm.ast.type.FieldType
-import io.github.charlietap.chasm.ast.type.StructType
 import io.github.charlietap.chasm.executor.runtime.error.InvocationError
 import io.github.charlietap.chasm.executor.runtime.exception.InvocationException
+import io.github.charlietap.chasm.ir.module.Index
+import io.github.charlietap.chasm.ir.type.FieldType
+import io.github.charlietap.chasm.ir.type.StructType
 
 fun StructType.field(
     index: Index.FieldIndex,
 ): FieldType = try {
-    fields[index.idx.toInt()]
+    fields[index.idx]
 } catch (_: IndexOutOfBoundsException) {
-    throw InvocationException(InvocationError.StructFieldLookupFailed(index.idx.toInt()))
+    throw InvocationException(InvocationError.StructFieldLookupFailed(index.idx))
 } catch (_: IllegalArgumentException) {
-    throw InvocationException(InvocationError.StructFieldLookupFailed(index.idx.toInt()))
+    throw InvocationException(InvocationError.StructFieldLookupFailed(index.idx))
 }
